requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
Flask-CORS==4.0.0